+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/verticalmode/CurrencySelectorOptionsFactory.kt
@@ -44,8 +44,8 @@ internal object CurrencySelectorOptionsFactory {
         }
 
         return CurrencySelectorOptions(
-            first = integrationCurrencyOption,
-            second = localCurrencyOption,
+            first = localCurrencyOption,
+            second = integrationCurrencyOption,
             selectedCode = selectedCode,
             exchangeRateText = exchangeRateText,
         )